requirements.txt
Django==2.2.12
django-tinymce==3.0.2
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
uWSGI==2.0.18
